Basic invalidation of the translation cache on memory writes
There is a cache of translations from instructions in memory to Scheme
code. When the instructions in memory change because of a write to
memory, these translations may also need to change. There is now a
quick and dirty table is used to mark translations as invalid when a
write is done in the same 128 byte area as an existing translation.
It's not quite correct yet, but works for a lot of code.