Faster memory and I/O hooking
There is now support for hooking all accesses to 4K pages, which
should be enough for emulating most memory mapped hardware. The
emulated RAM is constructed as a vector with one bytevector per page.
The page can now be a procedure instead of a bytevector. This approach
should have less overhead in the memory-*-{ref,set!} code than the
last one, since the bytevector type check is needed either way, and in
the case of a bytevector all code is inlined.

Additionally there is support for hooking I/O ports. The API is
exactly as the memory hooking API, except that the port size also
needs to be specified.