BIOS: Support writing to images, more floppy formats
Disk images can now be written to. (Write protect the image if you
don't want this). The size of floppy images is used as a heuristic to
guess the geometry, so more older floppies should be usable.

And DX is either #x00 or #x80 when booting, which turns out to be more
compatible.