Basic support for tracking undefined flags
There is now support for tracking which flags are undefined. This is
useful for the new test framework, which now masks away all undefined
flags before comparing the results.

This feature could also be used to track when an instruction (e.g. a
branch, SETcc or even an ADC/SBB instruction) uses an undefined flag,
which is interesting in itself. Depending on what the software is it
would indicate either a bug in the software or perhaps that a flag is
widely considered as defined for some instruction.