Update disassembly to work with the new machine-code API
The machine-code API has changed a little. The disassembly now has
proper addresses for branches. Also updated the CI tests to use the
build-scheme script from machine-code, to test with multiple
Schemes (and some are not working right now). The libraries have been
renamed to .sls, which seems more portable for the moment.